Update ohrstable.sh for the many renamed packages in ichorescent
Note that the Distribute Game menu no longer downloads ohrrpgce-player
packages from hamsterrepublic.com/dl/. So the only reason to update those
links is for linking to, which we don't seem to do yet.

Also, I changed the behaviour of updatelink() and fixed a bug where the 3rd
alternative prefix always failed.
If it finds ${OLDPREFIX[2]}${VERSION}${EXT} exists it creates a link from
${OLDPREFIX[2]}${EXT} rather than from ${PREFIX}${EXT}, to match the way the
obsolete symlinks in dl/ are actually set up: they link to the last version that
used a package with that naming scheme.

git-svn-id: https://rpg.hamsterrepublic.com/source/web@13290 7d344553-34f0-0310-a9b1-970ce8f1c3a2